+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+2.1.9
+----------
+`2024-08-12 · 1 🎉 `
+
+### 🎉 New Features
+- Feat: update session info listener handler
+
+2.1.8
+----------
+`2024-08-09 · 1 🎉 `
+
+### 🎉 New Features
+- Feat: manually send errors that occurs on whatsapp signup process
+
 2.1.7
 ----------
 `2024-07-29 · 1 🐛 `